+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -65,6 +65,7 @@ CHANGELOG
 6.3
 ---
 
+ * Add support for reading objects properties with `Unique` constraint `fields` option
  * Add method `getConstraint()` to `ConstraintViolationInterface`
  * Add `Uuid::TIME_BASED_VERSIONS` to match that a UUID being validated embeds a timestamp
  * Add the `pattern` parameter in violations of the `Regex` constraint